update scripts imputing functions from useful_functions
