+++ nutrition_labels/grant_data_processing.py
@@ -1,6 +1,8 @@
 import pandas as pd
 import re
 from bs4 import BeautifulSoup
+from nutrition_labels.useful_functions import remove_useless_string
+from nutrition_labels.usefull_functions import only_text
 
 # load data
 epmc_tags = pd.read_csv('data/raw/EPMC_relevant_tool_pubs_manual_edit.csv')
@@ -46,28 +48,6 @@ epmc_df = pd.DataFrame(epmc_list)
 epmc_df = epmc_df[epmc_df['code'].isin([1,2,3])] # Only selecting useful codes
 
 # getting WT grant number
-def remove_useless_string(string):
-    '''
-    cleans the grant descriptions of artifacts such as <br />
-    :param string: description string
-    :return: clean string
-    '''
-
-    soup = BeautifulSoup(string)
-    string_out = soup.get_text()
-    string_out = string_out.strip('\n')
-    string_out = string_out.strip('\xa0')
-    string_out = re.sub('  ','',string_out)
-    return(string_out)
-
-def only_text(string):
-    '''
-    removes non-alphanumeric characters and spaces to increase matching
-    :param string: description string
-    :return: clean string
-    '''
-    string = re.sub(' |\W','',string)
-    return(string)
 
 
 grant_ref = grant_data[['Grant Programme:Title','Internal ID','Description','Award Date']]
